requirements.txt
fastapi
uvicorn[standard]

# LangChain ecosystem
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma

# PDF loading
pypdf

# Web search (Wikipedia)
wikipedia

# Config & validation
python-dotenv
pydantic
